change around some variable names for consistency's sake:
  $current_key becomes $current_primary_key
  $current_fingerprinted_key becomes $current_key

This is because keys (both primary and sub) can have more than just
fingerprints following them.  The parser might want to know what was
the last key seen (either primary or sub) *and* it might want to know
the last primary key seen, since this will be associated with the sub
keys.