Remove effectively dead code for gpg2 checking
Two places attempted to test for gpg 1.4, or fall back to gpg2 if it
existed.  However, because it either 'die'd or 'last'ed, it was
impossible to reach the gpg2 iteration of the loop.  Simply remove the
useless loop, and test for the presence of 'gpg', as well as its
version.