Ensure that the trustdb is created before attempting to encrypt
gpg2 requires that $GNUPGHOME/trustdb.gpg exist, even for commands with
--trust-model=always which need not interact with the trust database.
gpg 1.4 does not have this restriction, and a fix for the gpg2 bug will
appear in 2.0.26; see [1].

This shows up as a test failure if using gpg2, but is also dependent on
test ordering; most other tests result in the creation of the trustdb
implicitly.  Unfortunately, as the format of the trustdb depends on the
version, it cannot simply be shipped as part of the distribution.

Explicitly create the trustdb before the one test that is susceptible to
the above gpg2 bug; all other tests currently deal gracefully with the
lack of a trustdb.

[1] http://git.gnupg.org/cgi-bin/gitweb.cgi?p=gnupg.git;a=commitdiff;h=07006c99