test suite: match plaintext output across versions of GnuPG
The human-readable version of --list-keys is *not* expected to be
static over time or as the user's environment changes (e.g. LANG or
LC_MESSAGES), so expecting it to be machine-parseable is probably a
mistake.

That said, some users might want to pull textual information about
specific keys to display directly to the user, so it's not a terrible
idea to have it in the test suite.

Modern GnuPG (2.1 or later) changes the default structure of the
human-readable output in a few significant ways:

 * it writes the path to the keyring as an absolute path, even if
   $GNUPGHOME is set to a non-absolute path.

 * it shows the calculated user id validity by default (see
   show-uid-validity in gpg's --list-options).  (note that this is a
   translated string, so that "unknown" (in the default C locale)
   becomes "inconnue" when LANG or LC_MESSAGES is set to fr_CH.UTF-8,
   for example.

 * it writes the key algorithm names differently (e.g. rsa2048 instead
   of 2048R)

 * it does not display the key ID at all by default

 * it displays the full fingerprint in compact form by default

This changeset fixes the test suite so that it can do a rough
verification of the human-readable text output by list_secret_keys in
the C locale in modern versions of GnuPG, while leaving it working for
older GnuPG suites.