Restore test resource file name construction with branch name
There are resource files under test/ and some of them have alternate
versions for modern GPG. Previously, I had updated the list_secret_keys
test to construct the filename with the system GPG's branch name
(classic/stable/modern). But the only versions of the file are named
with a '0' and 'modern'. This change restores the original functionality
using the new GPG version convenience methods in GnuPG::Interface.