Use GNUPGHOME for gpgconf instead of --homedir
GnuPG >= 2.1.0 starts gpg-agent for the gpg homedir if one is not
already running. gpgconf >= 2.1.0 can kill a gpg-agent for the homedir
with the --kill flag. But gpgconf >= 2.1.0 and <= 2.1.12 do not support
specifying homedir by the '--homedir' flag. That feature was introduced
in 2.1.13.

However, gpgconf >= 2.1.0 and <= 2.1.12 do respect the GNUPGHOME
environment variable. This change skips gpg-agent cleanup in the tests
for GnuPG < 2.1.0. It also updates the gpgconf call to use GNUPGHOME
instead of --homedir in the gpg-agent cleanup.